refactor: `model.py` へ移動
