+++ test/user_dict/test_user_dict_model.py
@@ -4,7 +4,7 @@ import pytest
 from pydantic import ValidationError
 
 from voicevox_engine.tts_pipeline.kana_converter import parse_kana
-from voicevox_engine.user_dict.part_of_speech_data import UserDictWord
+from voicevox_engine.user_dict.model import UserDictWord
 
 
 class TestModel(TypedDict):
