+++ voicevox_engine/app/routers/user_dict.py
@@ -6,12 +6,11 @@ from typing import Annotated
 from fastapi import APIRouter, Body, Depends, HTTPException, Path, Query
 from pydantic import ValidationError
 
+from voicevox_engine.user_dict.model import UserDictWord, WordTypes
 from voicevox_engine.user_dict.part_of_speech_data import (
     MAX_PRIORITY,
     MIN_PRIORITY,
     UserDictInputError,
-    UserDictWord,
-    WordTypes,
 )
 from voicevox_engine.user_dict.user_dict import UserDictionary
